requirements.txt
asgiref==3.5.0
Django==4.0.2
django-mysql==4.5.0
djangorestframework==3.13.1
mysqlclient==2.0.1
PyJWT==2.3.0
pytz==2021.3
sqlparse==0.4.2
